requirements.txt
fastapi
uvicorn
flask
flask-cors
python-dotenv
groq
numpy
sentence-transformers
faiss-cpu
